improve flow implementation

+++ src/main/kotlin/com/kotlincon/mvcdemo/Contact.kt
@@ -2,7 +2,7 @@ package com.kotlincon.mvcdemo
 
 import org.springframework.data.annotation.Id
 
-data class Contact(@Id var id: Long? = null,
+data class Contact(@Id var id: Int? = null,
                    val name: String,
                    val surname: String
 )
